use detect in qst.
this also changes the behaviour to chdir into the directory of the file
given, which may or may not be good. but i think it makes sense for a
lot of projects.

yet, we might want to cd to the *root* of the project sometimes, but
specify a file in it. not now, but probably should keep this in mind.